# git-contextor/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Git Contextor API",
    "version": "1.1.6",
    "description": "A comprehensive API for Git Contextor - a code context tool with vector search, AI chat, and real-time monitoring capabilities. Git Contextor indexes your codebase and provides semantic search, AI-powered chat, and code sharing features.",
    "contact": {
      "name": "Git Contextor Support",
      "url": "https://github.com/stromdao/git-contextor"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server (default port)"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "summary": "AI Chat",
        "description": "Sends a query to the AI using repository context to generate an answer.",
        "operationId": "chatQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language query for the AI",
                    "example": "Explain the auth flow"
                  },
                  "context_type": {
                    "type": "string",
                    "description": "Type of context to use for the query",
                    "example": "general"
                  },
                  "include_summary": {
                    "type": "boolean",
                    "description": "Whether to include a collection summary in the response",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string",
                      "example": "Explain the auth flow"
                    },
                    "response": {
                      "type": "string",
                      "description": "AI-generated response",
                      "example": "The authentication flow starts with a user providing credentials..."
                    },
                    "context": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResult"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/collection/summary": {
      "get": {
        "summary": "Get Collection Summary",
        "description": "Retrieves the AI-generated collection summary. If it doesn't exist, it will be generated on-demand.",
        "operationId": "getCollectionSummary",
        "responses": {
          "200": {
            "description": "Collection summary retrieved successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Markdown-formatted collection summary",
                  "example": "# Collection Summary\\n\\n## Cluster 1: Authentication & User Management\\nKey technologies: JWT, bcrypt, Express.js\\n..."
                }
              }
            }
          },
          "404": {
            "description": "Could not retrieve or create collection summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/summarize": {
      "post": {
        "summary": "Trigger Summary Generation",
        "description": "Manually triggers the generation of a new collection summary. This is an asynchronous operation.",
        "operationId": "triggerSummaryGeneration",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numClusters": {
                    "type": "integer",
                    "description": "Number of clusters to generate for the summary",
                    "example": 15
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Summary generation started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Collection summary generation started. This may take a few minutes."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/share/tunnel": {
      "post": {
        "summary": "Start Tunnel",
        "description": "Starts a tunnel service to expose the Git Contextor server publicly.",
        "operationId": "startTunnel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["service"],
                "properties": {
                  "service": {
                    "type": "string",
                    "description": "Tunnel service to use",
                    "example": "localtunnel"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Tunnel service is starting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tunnel service 'localtunnel' is starting."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Tunnel Status",
        "description": "Gets the current status of the tunnel service.",
        "operationId": "getTunnelStatus",
        "responses": {
          "200": {
            "description": "Tunnel status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["stopped", "starting", "running", "error"],
                      "example": "running"
                    },
                    "url": {
                      "type": "string",
                      "nullable": true,
                      "example": "https://abc123.localtunnel.me"
                    },
                    "service": {
                      "type": "string",
                      "nullable": true,
                      "example": "localtunnel"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Stop Tunnel",
        "description": "Stops the active tunnel service.",
        "operationId": "stopTunnel",
        "responses": {
          "200": {
            "description": "Tunnel stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tunnel stopped."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/shared/{shareId}/chat": {
      "post": {
        "summary": "Chat with Shared Repository",
        "description": "Sends a chat query to the AI using the shared repository context.",
        "operationId": "chatWithSharedRepo",
        "security": [
          {
            "ShareKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "description": "ID of the share",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language query for the AI"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "type": "object",
                      "properties": {
                        "query": {
                          "type": "string"
                        },
                        "response": {
                          "type": "string"
                        },
                        "context": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SearchResult"
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "share_id": {
                          "type": "string"
                        },
                        "queries_remaining": {
                          "type": "integer"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid share credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication. You can get this from the UI config endpoint or from your Git Contextor configuration."
      },
      "ShareKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-share-key",
        "description": "Share key for accessing shared repositories."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for MCP (Model Context Protocol) authentication. The token is the API key of the share."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "description": "Path to the file containing the result",
            "example": "src/auth/jwt.js"
          },
          "score": {
            "type": "number",
            "description": "Relevance score (0-1)",
            "example": 0.92
          },
          "content": {
            "type": "string",
            "description": "Content of the code chunk"
          },
          "startLine": {
            "type": "integer",
            "description": "Starting line number",
            "example": 10
          },
          "endLine": {
            "type": "integer",
            "description": "Ending line number",
            "example": 25
          }
        }
      },
...
    }
  },
...
}